requirements.txt
llama-cpp-python==0.2.56
numpy>=1.22,<1.25
matplotlib>=3.5.0,<3.8.0
tqdm>=4.65.0,<5.0.0 